#32 Correct imports in test file, refactor Chess Engine and add type guard for Piece
